feat: Add support for yarn on  Windows
Currently the package works incorrectly when Yarn is used to install
packages globally on Windows. Yarn uses different structure of placing
executables and content of packages. In order to handle it, just parse
the `.cmd` executable and search for `node_modules` in it. The file
should contain some special symbols describing the current directory
`%~dp0` followed by `node_modules` at some point. This is the real path
to the executable and it should always be correct. Replace the special
symbols with the real path and check if this is the package we are
looking for. Add additional check in the code to verify that the
returned directory contains package.json file with the same name that is
required by the caller.